+++ internal/testutils/jimmtest/utils.go
@@ -1,0 +1,20 @@
+// Copyright 2025 Canonical.
+
+package jimmtest
+
+import (
+	"errors"
+	"net"
+)
+
+// GetFreePort asks the kernel for a free open port that is ready to use.
+func GetFreePort() (int, error) {
+	if a, err := net.ResolveTCPAddr("tcp", "localhost:0"); err == nil {
+		var l *net.TCPListener
+		if l, err = net.ListenTCP("tcp", a); err == nil {
+			defer l.Close()
+			return l.Addr().(*net.TCPAddr).Port, nil
+		}
+	}
+	return 0, errors.New("Couldn't find any free port")
+}
